cookbook/examples/workflows/Deep-Research-Agent/requirements.txt
agno
openai
python-dotenv
arxiv
pypdf
wikipedia
ddgs
hackernews
pygithub
